Refactored half of auto tilemap.

+++ N8Engine/-Source-/Rendering/Window.cs
@@ -6,7 +6,7 @@ namespace N8Engine.Rendering
 {
     public static class Window
     {
-        internal const float RATIO_OF_HORIZONTAL_PIXELS_TO_VERTICAL_PIXELS = 2f;
+        internal const int RATIO_OF_HORIZONTAL_PIXELS_TO_VERTICAL_PIXELS = 2;
 
         private static readonly float _width = Console.WindowWidth;
         private static readonly float _height = Console.WindowHeight;
